Added functionality for Interrupt Handing
Off-loaded Interrupt Functions from main.c to dedicated .h/.c files. Added Timer for sensor read interrupts.

+++ Team205_FinalBoard.X/motor_control.c
@@ -2,7 +2,7 @@
 
 
 void motorController(uint8_t address, uint8_t speed, uint8_t dir){
-    uint8_t data = speed << 2 | dir;
+    uint8_t data = (uint8_t)(speed << 2 | dir);
     I2C1_Write1ByteRegister(address, MOTOR_CONTROL, data);
 }
